+++ ezcharts/plots/categorical.py
@@ -1,8 +1,10 @@
 """Categorical plots."""
 
-import numpy as np
+from bokeh.models import ColumnDataSource
+import bokeh.transform as bkt
+from seaborn.categorical import _BarPlotter
 
-from ezcharts.plots import Plot
+from ezcharts.plots import BokehPlot
 
 __all__ = [
     "catplot", "stripplot", "swarmplot", "boxplot", "violinplot",
@@ -48,55 +50,140 @@ def pointplot(*args, **kwargs):
 def barplot(
     data=None, *, x=None, y=None, hue=None, order=None, hue_order=None,
     estimator='mean', errorbar=('ci', 95), n_boot=1000, units=None, seed=None,
-    orient=None, color=None, palette=None, saturation=0.75, width=0.8,
+    orient=None, color=None, palette=None, saturation=1.0, width=0.8,
     errcolor='.26', errwidth=None, capsize=None, dodge=True, ci='deprecated',
-        ax=None, **kwargs):
-    """Show point estimates and confidence intervals as rectangular bars.
+    ax=None, **kwargs,
+):
+    """Show point estimates as rectangular bars.
 
     Contrary to the seaborn implementation, setting `dodge=False` does not
     result in overlaying the bars, but rather stacking them.
     """
-    x_name = x if x else data.columns[0]
-    y_name = y if y else data.columns[1]
+    palette = BokehPlot.colors if palette is None else palette
 
-    plt = Plot()
+    plotter = _BarPlotter(
+        x,
+        y,
+        hue,
+        data,
+        order,
+        hue_order,
+        estimator,
+        errorbar,
+        n_boot,
+        units,
+        seed,
+        orient,
+        color,
+        palette,
+        saturation,
+        width,
+        errcolor,
+        errwidth,
+        capsize,
+        dodge,
+    )
 
-    if hue not in data:
-        data = data[[x_name, y_name]]
+    if plotter.orient == "v":
+        plt = BokehPlot(x_range=plotter.group_names)
     else:
-        # Transform to [x, hue, hue, hue ...]
-        data = data.pivot(
-            columns=hue, values=y_name, index=x_name).reset_index(drop=False)
-        plt.legend = dict(orient='horizontal', top=25)
+        plt = BokehPlot(y_range=plotter.group_names)
 
-    if order:
-        mapper = {x: pos for (pos, x) in enumerate(order)}
-        data = data.sort_values(x_name, key=lambda x: x.map(mapper))
+    p = plt._fig
 
-    plt.xAxis = dict(
-        name=x_name,
-        type='category',
-        axisLabel=dict(interval=0, rotate=30),
-        axisTick=dict(alignWithLabel=True))
-
-    plt.yAxis = dict(name=y_name, type='value')
-
-    plt.add_dataset({
-        'id': 'raw',
-        'dimensions': data.columns,
-        'source': data.values})
-
-    # If there are nan values, replace with zeros
-    plt.dataset[0].source = np.nan_to_num(plt.dataset[0].source, nan=0)
-
-    # Automatically map series to columns in the dataset
-    if dodge:
-        plt.series = [{'type': 'bar'}] * (data.shape[1] - 1)
+    if plotter.plot_hues is None:
+        # simple barplot (i.e. only a single group of bars)
+        if plotter.orient == "v":
+            p.vbar(
+                x=plotter.group_names,
+                top=plotter.statistic,
+                width=0.9,
+                color=plotter.colors.as_hex(),
+                **kwargs,
+            )
+        else:
+            p.hbar(
+                y=plotter.group_names,
+                right=plotter.statistic,
+                height=0.9,
+                color=plotter.colors.as_hex(),
+                **kwargs,
+            )
     else:
-        plt.series = [{'type': 'bar', 'stack': x_name}] * (data.shape[1] - 1)
+        # groups of bars (either stacked or grouped)
+        data = dict(groups=plotter.group_names)
+        data.update(dict(zip(plotter.hue_names, plotter.statistic.T)))
+        data = ColumnDataSource(data)
 
-    # Show a label per bar by default as ezcharts 'auto' method is very conservative.
-    plt.xAxis.axisLabel = dict(interval=0)
+        if not dodge:
+            if plotter.orient == "v":
+                # stacked vertical bars
+                p.vbar_stack(
+                    plotter.hue_names,
+                    x="groups",
+                    width=0.95,
+                    color=plotter.colors.as_hex(),
+                    source=data,
+                    legend_label=plotter.hue_names,
+                    **kwargs
+                )
+            else:
+                # stacked horizontal bars
+                p.hbar_stack(
+                    plotter.hue_names,
+                    y="groups",
+                    height=0.95,
+                    color=plotter.colors.as_hex(),
+                    source=data,
+                    legend_label=plotter.hue_names,
+                    **kwargs
+                )
+        else:
+            # grouped plots (we need to add a series for each hue level)
+            for hue_level, offset, color in zip(
+                plotter.hue_names,
+                plotter.hue_offsets,
+                plotter.colors.as_hex(),
+            ):
+                plot_bars_kwargs = dict(
+                    source=data,
+                    color=color,
+                    line_color="white",
+                    legend_label=hue_level,
+                )
+                if plotter.orient == "v":
+                    plot_bars_func = p.vbar
+                    plot_bars_kwargs.update(
+                        dict(
+                            x=bkt.dodge("groups", offset, range=p.x_range),
+                            top=hue_level,
+                            width=plotter.nested_width,
+                        )
+                    )
+                else:
+                    plot_bars_func = p.hbar
+                    plot_bars_kwargs.update(
+                        dict(
+                            y=bkt.dodge("groups", offset, range=p.y_range),
+                            right=hue_level,
+                            height=plotter.nested_width,
+                        )
+                    )
+                plot_bars_func(**plot_bars_kwargs, **kwargs)
+
+        p.legend.orientation = "horizontal"
+        # even though we call `add_layout()` to pull the legend outside of the plot area
+        # below, we still need to set `location` to "top" here to make sure the legend
+        # is centered properly
+        p.legend.location = "top"
+        p.add_layout(p.legend[0], "above")
+
+    if plotter.orient == "v":
+        p.xgrid.grid_line_color = None
+        p.y_range.start = 0
+    else:
+        p.ygrid.grid_line_color = None
+        p.x_range.start = 0
 
     return plt